+++ poulpy-ckks/src/oep/polynomial_evaluation.rs
@@ -1,6 +1,7 @@
 use anyhow::Result;
 use poulpy_core::layouts::{
-    GGLWEInfos, GLWEToBackendMut, GLWEToBackendRef, SetBSGSMeta, prepared::GLWETensorKeyPreparedToBackendRef,
+    GGLWEInfos, GLWEInfos, GLWEToBackendMut, GLWEToBackendRef, SetBSGSMeta,
+    prepared::{GLWETensorKeyPrepared, GLWETensorKeyPreparedToBackendRef},
 };
 use poulpy_core::{
     GLWEAdd, GLWECopy, GLWEMulConst, GLWENormalize, GLWEPolynomialEvaluation, GLWEShift, GLWETensoring, GLWEZero,
@@ -11,9 +12,11 @@ use poulpy_hal::layouts::{Backend, Module, ScratchArena};
 
 use crate::{
     CKKSCtBounds, SetCKKSInfos,
-    api::{BSGSPolynomialInfos, CKKSAddOps, CKKSImagOps, CKKSMulAddOps, PowerBasisHelper},
+    api::{BSGSPolynomialInfos, CKKSAddOps, CKKSCopyOps, CKKSImagOps, CKKSMulAddOps, CKKSMulOps, CKKSSubOps, PowerBasisHelper},
     default::polynomial_evaluation::PolynomialEvaluationDefault,
-    layouts::CKKSModuleAlloc,
+    layouts::{CKKSCiphertext, CKKSModuleAlloc},
+    polynomial::{BSGSPolynomial, ComplexBSGSPolynomial},
+    power_basis::{PowerBasis, PowerBasisGen},
 };
 
 /// # Safety
@@ -37,30 +40,60 @@ pub unsafe trait CKKSPolynomialEvaluationImpl<BE: Backend>: Backend {
         G: PowerBasisHelper<BE, A>,
         T: GLWETensorKeyPreparedToBackendRef<BE> + GGLWEInfos;
 
-    #[allow(clippy::too_many_arguments)]
-    fn ckks_eval_poly_complex_const_coeffs_from_power_basis<R, B, A, G, T>(
+    fn ckks_eval_poly_complex_const_coeffs_from_power_basis<R, C, A, G, T>(
         module: &Module<BE>,
         res: &mut R,
-        poly_re: &B,
-        poly_im: &B,
+        poly: &ComplexBSGSPolynomial<C>,
         power_basis: &G,
         tsk: &T,
         scratch: &mut ScratchArena<'_, BE>,
     ) -> Result<()>
     where
         R: GLWEToBackendMut<BE> + CKKSCtBounds + SetCKKSInfos + SetBSGSMeta,
-        B: BSGSPolynomialInfos<BE>,
-        B::Coeffs: CKKSCtBounds,
+        C: GLWEToBackendRef<BE> + GLWEInfos + poulpy_core::layouts::BSGSMeta + CKKSCtBounds,
         A: GLWEToBackendRef<BE> + CKKSCtBounds + poulpy_core::layouts::BSGSMeta,
         G: PowerBasisHelper<BE, A>,
         T: GLWETensorKeyPreparedToBackendRef<BE> + GGLWEInfos;
+
+    fn ckks_eval_poly_real_const_coeffs<R, S, C>(
+        module: &Module<BE>,
+        dst: &mut R,
+        src: &S,
+        bsgs: &BSGSPolynomial<C>,
+        tsk: &GLWETensorKeyPrepared<BE::OwnedBuf, BE>,
+        scratch: &mut ScratchArena<'_, BE>,
+    ) -> Result<()>
+    where
+        R: GLWEToBackendMut<BE> + CKKSCtBounds + SetCKKSInfos + SetBSGSMeta,
+        S: GLWEToBackendRef<BE> + CKKSCtBounds,
+        C: GLWEToBackendRef<BE> + GLWEInfos + poulpy_core::layouts::BSGSMeta + CKKSCtBounds,
+        GLWETensorKeyPrepared<BE::OwnedBuf, BE>: GGLWEInfos + GLWETensorKeyPreparedToBackendRef<BE>,
+        CKKSCiphertext<BE::OwnedBuf>: GLWEToBackendMut<BE> + GLWEToBackendRef<BE> + CKKSCtBounds + SetCKKSInfos;
+
+    fn ckks_eval_poly_complex_const_coeffs<R, S, C>(
+        module: &Module<BE>,
+        dst: &mut R,
+        src: &S,
+        poly: &ComplexBSGSPolynomial<C>,
+        tsk: &GLWETensorKeyPrepared<BE::OwnedBuf, BE>,
+        scratch: &mut ScratchArena<'_, BE>,
+    ) -> Result<()>
+    where
+        R: GLWEToBackendMut<BE> + CKKSCtBounds + SetCKKSInfos + SetBSGSMeta,
+        S: GLWEToBackendRef<BE> + CKKSCtBounds,
+        C: GLWEToBackendRef<BE> + GLWEInfos + poulpy_core::layouts::BSGSMeta + CKKSCtBounds,
+        GLWETensorKeyPrepared<BE::OwnedBuf, BE>: GGLWEInfos + GLWETensorKeyPreparedToBackendRef<BE>,
+        CKKSCiphertext<BE::OwnedBuf>: GLWEToBackendMut<BE> + GLWEToBackendRef<BE> + CKKSCtBounds + SetCKKSInfos;
 }
 
 unsafe impl<BE: Backend> CKKSPolynomialEvaluationImpl<BE> for BE
 where
     Module<BE>: GiantStepTensorBounds<BE>
         + CKKSAddOps<BE>
+        + CKKSCopyOps<BE>
         + CKKSImagOps<BE>
+        + CKKSMulOps<BE>
+        + CKKSSubOps<BE>
         + CKKSMulAddOps<BE>
         + GLWEMulConst<BE>
         + GLWEAdd<BE>
@@ -93,28 +126,75 @@ where
         module.ckks_eval_poly_real_const_coeffs_from_power_basis_default::<R, B, A, G, T>(res, poly, power_basis, tsk, scratch)
     }
 
-    fn ckks_eval_poly_complex_const_coeffs_from_power_basis<R, B, A, G, T>(
+    fn ckks_eval_poly_complex_const_coeffs_from_power_basis<R, C, A, G, T>(
         module: &Module<BE>,
         res: &mut R,
-        poly_re: &B,
-        poly_im: &B,
+        poly: &ComplexBSGSPolynomial<C>,
         power_basis: &G,
         tsk: &T,
         scratch: &mut ScratchArena<'_, BE>,
     ) -> Result<()>
     where
         R: GLWEToBackendMut<BE> + CKKSCtBounds + SetCKKSInfos + SetBSGSMeta,
-        B: BSGSPolynomialInfos<BE>,
-        B::Coeffs: CKKSCtBounds,
+        C: GLWEToBackendRef<BE> + GLWEInfos + poulpy_core::layouts::BSGSMeta + CKKSCtBounds,
         A: GLWEToBackendRef<BE> + CKKSCtBounds + poulpy_core::layouts::BSGSMeta,
         G: PowerBasisHelper<BE, A>,
         T: GLWETensorKeyPreparedToBackendRef<BE> + GGLWEInfos,
     {
-        module.ckks_eval_poly_complex_const_coeffs_from_power_basis_default::<R, B, A, G, T>(
-            res,
-            poly_re,
-            poly_im,
-            power_basis,
+        module.ckks_eval_poly_complex_const_coeffs_from_power_basis_default::<R, C, A, G, T>(res, poly, power_basis, tsk, scratch)
+    }
+
+    fn ckks_eval_poly_real_const_coeffs<R, S, C>(
+        module: &Module<BE>,
+        dst: &mut R,
+        src: &S,
+        bsgs: &BSGSPolynomial<C>,
+        tsk: &GLWETensorKeyPrepared<BE::OwnedBuf, BE>,
+        scratch: &mut ScratchArena<'_, BE>,
+    ) -> Result<()>
+    where
+        R: GLWEToBackendMut<BE> + CKKSCtBounds + SetCKKSInfos + SetBSGSMeta,
+        S: GLWEToBackendRef<BE> + CKKSCtBounds,
+        C: GLWEToBackendRef<BE> + GLWEInfos + poulpy_core::layouts::BSGSMeta + CKKSCtBounds,
+        GLWETensorKeyPrepared<BE::OwnedBuf, BE>: GGLWEInfos + GLWETensorKeyPreparedToBackendRef<BE>,
+        CKKSCiphertext<BE::OwnedBuf>: GLWEToBackendMut<BE> + GLWEToBackendRef<BE> + CKKSCtBounds + SetCKKSInfos,
+    {
+        let mut x1 = module.ckks_ciphertext_alloc_from_infos(src);
+        module.ckks_copy(&mut x1, src, scratch)?;
+        let mut power_basis = PowerBasis::new(bsgs.basis(), x1);
+        power_basis.populate(bsgs.degree(), bsgs.log_split(), bsgs.parity(), module, tsk, scratch)?;
+        module.ckks_eval_poly_real_const_coeffs_from_power_basis_default::<R, BSGSPolynomial<C>, CKKSCiphertext<BE::OwnedBuf>, PowerBasis<CKKSCiphertext<BE::OwnedBuf>>, GLWETensorKeyPrepared<BE::OwnedBuf, BE>>(
+            dst,
+            bsgs,
+            &power_basis,
+            tsk,
+            scratch,
+        )
+    }
+
+    fn ckks_eval_poly_complex_const_coeffs<R, S, C>(
+        module: &Module<BE>,
+        dst: &mut R,
+        src: &S,
+        poly: &ComplexBSGSPolynomial<C>,
+        tsk: &GLWETensorKeyPrepared<BE::OwnedBuf, BE>,
+        scratch: &mut ScratchArena<'_, BE>,
+    ) -> Result<()>
+    where
+        R: GLWEToBackendMut<BE> + CKKSCtBounds + SetCKKSInfos + SetBSGSMeta,
+        S: GLWEToBackendRef<BE> + CKKSCtBounds,
+        C: GLWEToBackendRef<BE> + GLWEInfos + poulpy_core::layouts::BSGSMeta + CKKSCtBounds,
+        GLWETensorKeyPrepared<BE::OwnedBuf, BE>: GGLWEInfos + GLWETensorKeyPreparedToBackendRef<BE>,
+        CKKSCiphertext<BE::OwnedBuf>: GLWEToBackendMut<BE> + GLWEToBackendRef<BE> + CKKSCtBounds + SetCKKSInfos,
+    {
+        let mut x1 = module.ckks_ciphertext_alloc_from_infos(src);
+        module.ckks_copy(&mut x1, src, scratch)?;
+        let mut power_basis = PowerBasis::new(poly.re.basis(), x1);
+        power_basis.populate(poly.re.degree(), poly.re.log_split(), poly.re.parity(), module, tsk, scratch)?;
+        module.ckks_eval_poly_complex_const_coeffs_from_power_basis_default::<R, C, CKKSCiphertext<BE::OwnedBuf>, PowerBasis<CKKSCiphertext<BE::OwnedBuf>>, GLWETensorKeyPrepared<BE::OwnedBuf, BE>>(
+            dst,
+            poly,
+            &power_basis,
             tsk,
             scratch,
         )